Change license to MIT
It's MIT now, have fun. I'm not discontinuing it or anything, just wanted to make it so if anyone needs to copy code from it they can. Depending on how much you copy you may be technically inclined to include this license with your product, but that license only would need to apply to the parts you copy, not your whole project (unlike how it was with the GPL). Plus I probably wont sue you anyway.